+++ pkg/clientset/typed/configuration/v1alpha1/configuration_client.go
@@ -29,6 +29,7 @@ import (
 type ConfigurationV1alpha1Interface interface {
 	RESTClient() rest.Interface
 	IngressClassParametersesGetter
+	KongCustomEntitiesGetter
 	KongLicensesGetter
 	KongVaultsGetter
 }
@@ -40,6 +41,10 @@ type ConfigurationV1alpha1Client struct {
 
 func (c *ConfigurationV1alpha1Client) IngressClassParameterses(namespace string) IngressClassParametersInterface {
 	return newIngressClassParameterses(c, namespace)
+}
+
+func (c *ConfigurationV1alpha1Client) KongCustomEntities(namespace string) KongCustomEntityInterface {
+	return newKongCustomEntities(c, namespace)
 }
 
 func (c *ConfigurationV1alpha1Client) KongLicenses() KongLicenseInterface {