+++ CHANGELOG.md
@@ -130,9 +130,13 @@ Adding a new version? You'll need three changes:
   for headers specified with `konghq.com/headers.*` annotations. Moreover parsing a content of `konghq.com/headers.*`
   is more robust - leading and trailing whitespace characters are discarded.
   [#5977](https://github.com/Kong/kubernetes-ingress-controller/pull/5977)
+- Added the CRD `KongCustomEntity` to support custom Kong entities that are not
+  defined in KIC yet.
+  [#5982](https://github.com/Kong/kubernetes-ingress-controller/pull/5982)
 - Added `FallbackConfiguration` feature gate to enable the controller to generate a fallback configuration
   for Kong when it fails to apply the configuration. The feature gate is disabled by default.
   [#5993](https://github.com/Kong/kubernetes-ingress-controller/pull/5993)
+
 
 ### Fixed